+++ crawler/hdu/hdu.py
@@ -431,27 +431,3 @@ class HDUCrawler(BaseCrawler):
                 )
                 if stop_fetching:
                     break
-
-
-if __name__ == "__main__":
-    crawler = HDUCrawler()
-    crawler.log("info", "HDU Crawler is disabled.")
-    crawler.deinit_driver()
-    raise RuntimeError("HDU Crawler is disabled.")
-
-    # try:
-    #     crawler.log(
-    #         "important", "hdu Crawler started at " + datetime.now(beijing).isoformat()
-    #     )
-    #     crawler.fetch_contests()
-    #     crawler.log("info", "Contests fetched successfully.")
-    #     crawler.fetch_submissions()
-    #     crawler.log("info", "Submissions fetched successfully.")
-    #     crawler.log(
-    #         "important",
-    #         "hdu Crawler finished successfully at " + datetime.now(beijing).isoformat(),
-    #     )
-    #     crawler.finish()
-    # except Exception as e:
-    #     crawler.log("fatal", f"An error occurred: {e}")
-    #     crawler.deinit_driver()